requirements.txt
langchain
langchain-groq
langchain-community
langchain-core
faiss-cpu
python-dotenv
pypdf
tiktoken